refactor: split backend.rs into focused route-group modules (#202)
Mechanically split the 3,771-line server/src/backend.rs into a thin module
root (router + AppState + shared core + test_support fixtures) plus ten
route-group submodules under server/src/backend/ (health, settings, ebooks,
search, covers, overrides, authors, author_photos, series, tags). Tests move
next to the handlers they exercise; cross-cutting router tests stay in the
parent. No behavior change — public API (rest_router, AppState, build_id,
rate-limit constants) is preserved via the module root.
