+++ server/src/backend/ebooks.rs
@@ -1,0 +1,374 @@
+use axum::{
+    extract::{Path, State},
+    response::{IntoResponse, Response},
+    Json,
+};
+use omnibus_db::{self as db, scanner};
+
+use super::{internal, with_pagination_headers, AppState};
+use crate::auth::AuthUser;
+
+pub(super) async fn get_ebooks(_user: AuthUser, State(state): State<AppState>) -> Response {
+    let settings = match db::get_settings(&state.pool).await {
+        Ok(s) => s,
+        Err(error) => return internal("read settings", error),
+    };
+    match db::library_from_db_with_total(&state.pool, settings.ebook_library_path.as_deref()).await
+    {
+        Ok((library, total)) => with_pagination_headers(Json(library).into_response(), total),
+        Err(error) => internal("read books", error),
+    }
+}
+
+pub(super) async fn get_ebook_by_uuid(
+    _user: AuthUser,
+    State(state): State<AppState>,
+    Path(uuid): Path<String>,
+) -> Response {
+    match db::get_book_by_uuid(&state.pool, &uuid).await {
+        Ok(Some(book)) => Json(book).into_response(),
+        Ok(None) => axum::http::StatusCode::NOT_FOUND.into_response(),
+        Err(error) => internal("read book", error),
+    }
+}
+
+pub(super) async fn get_library(_user: AuthUser, State(state): State<AppState>) -> Response {
+    match db::get_settings(&state.pool).await {
+        Ok(settings) => {
+            let contents = scanner::scan_libraries(
+                settings.ebook_library_path.as_deref(),
+                settings.audiobook_library_path.as_deref(),
+            );
+            Json(contents).into_response()
+        }
+        Err(error) => internal("read settings", error),
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use axum::{body::to_bytes, http::StatusCode};
+    use omnibus_shared::Settings;
+    use tower::ServiceExt;
+
+    use super::*;
+    use crate::auth::test_support as auth_test_support;
+    use crate::backend::test_support::*;
+
+    #[tokio::test]
+    async fn api_get_library_returns_empty_sections_when_paths_not_configured() {
+        let (app, _state, pool) = fixture().await;
+        let user = auth_test_support::create_user(&pool, "alice").await;
+        let token = auth_test_support::bearer_token(&pool, user.id).await;
+
+        let response = app
+            .oneshot(get_with_bearer("/api/library", &token))
+            .await
+            .expect("request should succeed");
+        assert_eq!(response.status(), StatusCode::OK);
+
+        let bytes = to_bytes(response.into_body(), usize::MAX).await.unwrap();
+        let contents: omnibus_shared::LibraryContents = serde_json::from_slice(&bytes).unwrap();
+        assert!(contents.ebooks.path.is_none());
+        assert_eq!(contents.ebooks.total_files, 0);
+        assert!(contents.audiobooks.path.is_none());
+        assert_eq!(contents.audiobooks.total_files, 0);
+    }
+
+    #[tokio::test]
+    async fn api_get_library_reports_error_for_nonexistent_path() {
+        let (_, _, pool) = fixture().await;
+        db::set_settings(
+            &pool,
+            &Settings {
+                ebook_library_path: Some("/does/not/exist/omnibus_test".to_string()),
+                audiobook_library_path: None,
+            },
+        )
+        .await
+        .expect("set should succeed");
+        let user = auth_test_support::create_user(&pool, "alice").await;
+        let token = auth_test_support::bearer_token(&pool, user.id).await;
+        let app = crate::backend::rest_router(AppState::new(pool));
+
+        let response = app
+            .oneshot(get_with_bearer("/api/library", &token))
+            .await
+            .expect("request should succeed");
+        assert_eq!(response.status(), StatusCode::OK);
+
+        let bytes = to_bytes(response.into_body(), usize::MAX).await.unwrap();
+        let contents: omnibus_shared::LibraryContents = serde_json::from_slice(&bytes).unwrap();
+        assert!(contents.ebooks.error.is_some());
+        assert!(contents.audiobooks.path.is_none());
+    }
+
+    #[tokio::test]
+    async fn api_get_ebooks_returns_empty_when_path_not_configured() {
+        let (app, _state, pool) = fixture().await;
+        let user = auth_test_support::create_user(&pool, "alice").await;
+        let token = auth_test_support::bearer_token(&pool, user.id).await;
+
+        let response = app
+            .oneshot(get_with_bearer("/api/ebooks", &token))
+            .await
+            .expect("request should succeed");
+        assert_eq!(response.status(), StatusCode::OK);
+
+        let bytes = to_bytes(response.into_body(), usize::MAX).await.unwrap();
+        let lib: omnibus_shared::EbookLibrary = serde_json::from_slice(&bytes).unwrap();
+        assert!(lib.path.is_none());
+        assert!(lib.books.is_empty());
+        assert!(lib.error.is_none());
+    }
+
+    #[tokio::test]
+    async fn api_get_ebooks_returns_empty_library_for_configured_path_without_index() {
+        // /api/ebooks now reads from the books table; an unindexed path
+        // surfaces as an empty library at that path, not an error.
+        let pool = db::init_db("sqlite::memory:")
+            .await
+            .expect("db should initialize");
+        let path = "/does/not/exist/omnibus_ebook_test";
+        db::set_settings(
+            &pool,
+            &Settings {
+                ebook_library_path: Some(path.to_string()),
+                audiobook_library_path: None,
+            },
+        )
+        .await
+        .expect("set should succeed");
+        let user = auth_test_support::create_user(&pool, "alice").await;
+        let token = auth_test_support::bearer_token(&pool, user.id).await;
+        let app = crate::backend::rest_router(AppState::new(pool));
+
+        let response = app
+            .oneshot(get_with_bearer("/api/ebooks", &token))
+            .await
+            .expect("request should succeed");
+        assert_eq!(response.status(), StatusCode::OK);
+
+        let bytes = to_bytes(response.into_body(), usize::MAX).await.unwrap();
+        let lib: omnibus_shared::EbookLibrary = serde_json::from_slice(&bytes).unwrap();
+        assert_eq!(lib.path.as_deref(), Some(path));
+        assert!(lib.books.is_empty());
+        assert!(lib.error.is_none());
+    }
+
+    #[tokio::test]
+    async fn api_get_ebooks_sets_total_count_header_with_indexed_library() {
+        // Issue #81: every /api/ebooks response carries an X-Total-Count
+        // header. When the result fits under MAX_BOOKS_RETURNED no
+        // X-Total-Cap header is set, so the client knows the response
+        // is complete.
+        let (app, _state, pool) = fixture().await;
+        let user = auth_test_support::create_user(&pool, "alice").await;
+        let token = auth_test_support::bearer_token(&pool, user.id).await;
+
+        db::set_settings(
+            &pool,
+            &Settings {
+                ebook_library_path: Some("/lib".into()),
+                audiobook_library_path: None,
+            },
+        )
+        .await
+        .unwrap();
+        db::replace_books(
+            &pool,
+            "/lib",
+            vec![
+                db::ebook::IndexedBook {
+                    metadata: omnibus_shared::EbookMetadata {
+                        filename: "alpha.epub".into(),
+                        title: Some("A".into()),
+                        ..Default::default()
+                    },
+                    cover: None,
+                    mtime_epoch: 0,
+                    size_bytes: 0,
+                },
+                db::ebook::IndexedBook {
+                    metadata: omnibus_shared::EbookMetadata {
+                        filename: "beta.epub".into(),
+                        title: Some("B".into()),
+                        ..Default::default()
+                    },
+                    cover: None,
+                    mtime_epoch: 0,
+                    size_bytes: 0,
+                },
+            ],
+        )
+        .await
+        .unwrap();
+
+        let response = app
+            .oneshot(get_with_bearer("/api/ebooks", &token))
+            .await
+            .expect("request should succeed");
+        assert_eq!(response.status(), StatusCode::OK);
+        assert_eq!(
+            response
+                .headers()
+                .get("X-Total-Count")
+                .and_then(|v| v.to_str().ok()),
+            Some("2"),
+            "X-Total-Count must reflect the true row count"
+        );
+        assert!(
+            response.headers().get("X-Total-Cap").is_none(),
+            "X-Total-Cap must not be set when the response is not truncated"
+        );
+    }
+
+    #[tokio::test]
+    async fn api_get_ebooks_sets_total_cap_header_when_truncated() {
+        // Issue #81: when the underlying row count exceeds MAX_BOOKS_RETURNED,
+        // the response body is capped and X-Total-Cap is attached so the
+        // client knows the JSON it received isn't the full set.
+        let (app, _state, pool) = fixture().await;
+        let user = auth_test_support::create_user(&pool, "alice").await;
+        let token = auth_test_support::bearer_token(&pool, user.id).await;
+
+        db::set_settings(
+            &pool,
+            &Settings {
+                ebook_library_path: Some("/lib".into()),
+                audiobook_library_path: None,
+            },
+        )
+        .await
+        .unwrap();
+
+        // Bulk-seed > MAX_BOOKS_RETURNED rows directly so the test runtime
+        // stays in milliseconds. The cap behavior only needs rows to exist;
+        // the indexer's full m2m wiring isn't relevant here.
+        let total = db::MAX_BOOKS_RETURNED + 5;
+        let lib_id: i64 = sqlx::query_scalar(
+            "INSERT INTO libraries (path, display_name) VALUES ('/lib', 'lib') RETURNING id",
+        )
+        .fetch_one(&pool)
+        .await
+        .unwrap();
+        sqlx::query(
+            r#"
+            WITH RECURSIVE n(i) AS (
+                SELECT 1
+                UNION ALL
+                SELECT i + 1 FROM n WHERE i < ?
+            )
+            INSERT INTO books (uuid, library_id, path, title, sort)
+            SELECT 'uuid-' || i, ?, '/lib/b' || i, 'Title ' || i,
+                   'Title ' || printf('%010d', i)
+              FROM n
+            "#,
+        )
+        .bind(total)
+        .bind(lib_id)
+        .execute(&pool)
+        .await
+        .unwrap();
+
+        let response = app
+            .oneshot(get_with_bearer("/api/ebooks", &token))
+            .await
+            .expect("request should succeed");
+        assert_eq!(response.status(), StatusCode::OK);
+        assert_eq!(
+            response
+                .headers()
+                .get("X-Total-Count")
+                .and_then(|v| v.to_str().ok())
+                .and_then(|s| s.parse::<i64>().ok()),
+            Some(total),
+            "X-Total-Count must report the uncapped row count"
+        );
+        assert_eq!(
+            response
+                .headers()
+                .get("X-Total-Cap")
+                .and_then(|v| v.to_str().ok())
+                .and_then(|s| s.parse::<i64>().ok()),
+            Some(db::MAX_BOOKS_RETURNED),
+            "X-Total-Cap must equal MAX_BOOKS_RETURNED when the response is truncated"
+        );
+    }
+
+    #[tokio::test]
+    async fn api_get_ebook_returns_200_with_metadata() {
+        let (app, _state, pool) = fixture().await;
+        let user = auth_test_support::create_user(&pool, "alice").await;
+        let token = auth_test_support::bearer_token(&pool, user.id).await;
+
+        db::replace_books(
+            &pool,
+            "/lib",
+            vec![db::ebook::IndexedBook {
+                metadata: omnibus_shared::EbookMetadata {
+                    filename: "alpha.epub".into(),
+                    title: Some("Alpha Book".into()),
+                    ..Default::default()
+                },
+                cover: None,
+                mtime_epoch: 0,
+                size_bytes: 0,
+            }],
+        )
+        .await
+        .unwrap();
+
+        let books = db::list_books(&pool, "/lib").await.unwrap();
+        let id = books[0].id;
+        let uuid = books[0].unique_identifier.clone().unwrap();
+
+        let response = app
+            .oneshot(get_with_bearer(&format!("/api/ebooks/{uuid}"), &token))
+            .await
+            .expect("request should succeed");
+        assert_eq!(response.status(), StatusCode::OK);
+
+        let bytes = to_bytes(response.into_body(), usize::MAX).await.unwrap();
+        let book: omnibus_shared::EbookMetadata = serde_json::from_slice(&bytes).unwrap();
+        assert_eq!(book.title.as_deref(), Some("Alpha Book"));
+        assert_eq!(book.id, id);
+    }
+
+    #[tokio::test]
+    async fn api_get_ebook_returns_404_for_unknown_id() {
+        let (app, _state, pool) = fixture().await;
+        let user = auth_test_support::create_user(&pool, "alice").await;
+        let token = auth_test_support::bearer_token(&pool, user.id).await;
+
+        let response = app
+            .oneshot(get_with_bearer("/api/ebooks/9999", &token))
+            .await
+            .expect("request should succeed");
+        assert_eq!(response.status(), StatusCode::NOT_FOUND);
+    }
+
+    #[tokio::test]
+    async fn api_get_ebook_returns_401_when_anonymous() {
+        let (app, _state, _pool) = fixture().await;
+        let response = app
+            .oneshot(get_anon("/api/ebooks/1"))
+            .await
+            .expect("request should succeed");
+        assert_eq!(response.status(), StatusCode::UNAUTHORIZED);
+    }
+
+    #[tokio::test]
+    async fn api_library_returns_401_when_anonymous() {
+        let (app, _, _) = fixture().await;
+        let res = app.oneshot(get_anon("/api/library")).await.unwrap();
+        assert_eq!(res.status(), StatusCode::UNAUTHORIZED);
+    }
+
+    #[tokio::test]
+    async fn api_ebooks_returns_401_when_anonymous() {
+        let (app, _, _) = fixture().await;
+        let res = app.oneshot(get_anon("/api/ebooks")).await.unwrap();
+        assert_eq!(res.status(), StatusCode::UNAUTHORIZED);
+    }
+}
